requirements.txt
# Web Framework
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Math & Symbolic
sympy==1.12
numpy==1.24.3
scikit-learn==1.3.2

# OCR & Image Processing
pytesseract==0.3.10
pillow==10.1.0
opencv-python==4.8.1.78

# LLM API
google-generativeai==0.3.0

# Mathematical Verification (HuggingFace)
math-verify>=0.8.0
antlr4-python3-runtime>=4.9.3,<=4.13.2
datasets>=2.14.0

# Handwritten Math OCR
torch>=1.9.0
tqdm>=4.50.0
editdistance

# Utilities
requests==2.31.0
pydantic==2.4.2
python-dotenv==1.0.0
